pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["sat"]
readme = "README.md"
packages = [{include = "test_task"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
sqlalchemy = "^2.0.13"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
aiohttp = "^3.8.4"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
